added missing files to unstage
