+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "quake",
   "version": "1.0.0",
-  "description": "",
+  "description": "Index pulls from the last day of 2.5+mag quakes.  must have statenames.js in the same directory as index.html # To-do",
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon ./server/server.js"
+    "start": "nodemon server/server.js"
   },
   "repository": {
     "type": "git",
@@ -23,9 +23,11 @@
     "datamaps": "^0.4.2",
     "express": "^4.13.3",
     "nodemon": "^1.8.1",
+    "npm": "^3.5.3",
     "path": "^0.12.7",
     "request": "^2.67.0",
     "require": "^2.4.20",
     "topojson": "^1.6.20"
-  }
+  },
+  "devDependencies": {}
 }
